refactor: addPostToList function added to utils (#12)
