+++ index.html
@@ -761,4 +761,5 @@
     ></script>
     <script src="/js/main.js" type="module"></script>
     <script src="./js/login.js"></script>
+    <script src="./js/utils.js" type="module"></script>
 </html>
