+++ js/main.js
@@ -1,19 +1,6 @@
 import { getAllPosts } from './utils.js'
-import { createPostElement } from './utils.js'
+import { addPostsToList } from './utils.js'
 
 const posts = await getAllPosts()
-const postsList = document.getElementById('postsList')
-const locationToInsertFirstPostImage = document.getElementById('firstPostImage')
-postsList.innerHTML = ''
 
-const addPostToList = (postData) => {
-    const firstPostImage = postData[0].imageValue
-    locationToInsertFirstPostImage.src = firstPostImage
-
-    postData.forEach((post) => {
-        const postElement = createPostElement(post)
-        postsList.appendChild(postElement)
-    })
-}
-
-addPostToList(posts)
+addPostsToList(posts)
